html/browsers/browsing-the-web/navigating-across-documents/014.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
    steps: [12]
  - id: browsing-the-web.html#javascript-protocol
  - id: text-level-semantics.html#the-a-element
  - id: webappapis.html#the-event-handler-processing-algorithm
  - id: form-control-infrastructure.html#concept-form-submit
  - id: form-control-infrastructure.html#submit-get-action
  - id: form-control-infrastructure.html#plan-to-navigate
  - id: form-control-infrastructure.html#planned-navigation
  - id: iframe-embed-object.html#the-iframe-element:about:blank

description: |
  1. This test creates a form with `target="test"` and an anchor with both an
      onclick and href.
  2. The test "clicks" the anchor, which will trigger the onclick handler
      operation, which will attempt to submit the form. The form's `action` is
      a `javascript:` url, which must be evaluated and executed asynchronously.
  3. location.href is used to prove that the `javascript:` url's contents are
      executed in the iframe.

/-->
<title>Link with onclick form submit to javascript: url, order of execution</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<form target="test" action="javascript:parent.reports.push(location.href);"></form>
<a target="test" onclick="document.forms[0].submit()">Test</a>
<script>
var t = async_test(undefined, {timeout:4000});
var reports = [];
t.step(function() {
  document.getElementsByTagName("a")[0].click();
  reports.push("after click");
});
onload = t.step_func(function() {
  // javascript: executions are async.
  assert_array_equals(reports, ["after click", "about:blank"]);
  t.done();
});
</script>
